requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
apscheduler==3.7.0
asgiref==3.3.4
autopep8==1.5.7
babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.0
coreapi==2.3.3
coreschema==0.0.4
django-admin-honeypot==1.1.0
django-apscheduler==0.6.0
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
django-filter==2.4.0
django-phonenumber-field[phonenumberslite]==5.2.0
django-resized==0.3.11
django==3.2.3
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
jinja2==3.0.1
markdown==3.3.4; python_version >= '3.6'
markupsafe==2.0.1
martor==1.6.4
packaging==21.0; python_version >= '3.6'
phonenumberslite==8.12.31
pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyjwt==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
requests==2.26.0
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.10
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
